Extract link component from editable text
REDMINE-20673
